v2 crash brought down webform.  Don't truncate tables unless we had a valid import

+++ imports/import_v2_drugs.php
@@ -44,6 +44,9 @@ function import_v2_drugs() {
 
     $vals[] = '('.implode(', ', $val).')';
   }
+
+  if ( ! count($vals)) return log_error('No v2 Drugs to Import', get_defined_vars());
+  
   //Replace Staging Table with New Data
   $mysql->run('TRUNCATE TABLE gp_drugs_v2');
